requirements.txt
fastapi
uvicorn
openai
pillow
scikit-learn
numpy
python-multipart
requests
gitpython
duckdb
beautifulsoup4
openai-whisper
markdown
pandas
nodeenv
